Moved simulator_tvb from tvb_epilepsy/tvb_api to tvb_epilepsy/service/simulator, updated usages and deleted old tvb_api.py.
