+++ tvb_epilepsy/service/pse/simulation_pse_service.py
@@ -6,7 +6,7 @@ from tvb_epilepsy.base.utils.data_structures_utils import formal_repr
 from tvb_epilepsy.service.epileptor_model_factory import model_build_dict
 from tvb_epilepsy.service.pse.pse_service import ABCPSEService
 from tvb_epilepsy.base.model.disease_hypothesis import DiseaseHypothesis
-from tvb_epilepsy.tvb_api.simulator_tvb import SimulatorTVB
+from tvb_epilepsy.service.simulator.simulator_tvb import SimulatorTVB
 from tvb_epilepsy.service.simulator.custom_simulator import custom_model_builder
